Fixed GetMatch includeTimeline parameter parsing issue (#409)

+++ RiotSharpTest/RiotApiTestBase.cs
@@ -8,13 +8,13 @@ namespace RiotSharpTest
 {
     class RiotApiTestBase : CommonTestBase
     {
-        public static int gameId = int.Parse(ConfigurationManager.AppSettings["GameId"]);
+        public static long gameId = long.Parse(ConfigurationManager.AppSettings["GameId"]);
         public static int championId = int.Parse(ConfigurationManager.AppSettings["ChampionId"]);
         public static Platform summoner1Platform = (Platform)Enum.Parse(typeof(Platform), 
             ConfigurationManager.AppSettings["Summoner1Platform"]);
         public static int summoner1MasteryChampionId = int.Parse(ConfigurationManager.AppSettings["Summoner1MasteryChampionId"]);
         public static int summoner1MasteryChampionLevel = int.Parse(ConfigurationManager.AppSettings["Summoner1MasteryChampionLevel"]);
-        public static int unrankedSummonerId = int.Parse(ConfigurationManager.AppSettings["UnrankedSummonerId"]);
+        public static long unrankedSummonerId = long.Parse(ConfigurationManager.AppSettings["UnrankedSummonerId"]);
         public static List<long> summonerIds = ConfigurationManager.AppSettings["SummonerIds"].Split(',').Select(long.Parse).ToList();
         public static List<string> summonerNames = 
             ConfigurationManager.AppSettings["SummonerNames"].Split(',').Select(name => name.Trim()).ToList();
